+++ bank-account/src/main/java/com/dac/bank_account/command/dto/response/MovementResponseDTO.java
@@ -1,12 +1,8 @@
 package com.dac.bank_account.command.dto.response;
 
-
-
-import java.math.BigDecimal;
-
 public record MovementResponseDTO(
         String conta,
         String data,
-        BigDecimal saldo
+        Double saldo
 ) {
 }
