+++ bank-account/src/main/java/com/dac/bank_account/command/dto/response/TransferResponseDTO.java
@@ -1,13 +1,10 @@
 package com.dac.bank_account.command.dto.response;
 
-
-import java.math.BigDecimal;
-
 public record TransferResponseDTO(
         String conta,
         String data,
         String destino,
-        BigDecimal saldo,
-        BigDecimal valor
+        Double saldo,
+        Double valor
 ) {
 }
